restructure /test, move gen.py to /rdocx
