+++ test/README_test.md
@@ -1,9 +1,6 @@
 # Unittest
-## test_rsid(pytest).py only works on Windows cmd
-## test_rsid(pytest).py relies on gen.py to generate edited version of .docx files
-## gen.py use pywin32 python module to manipulate .docx file
-## pywin32 module calls MS word to perform operations (eg: edit, create) on a .docx file
-## process is slow because it is calling MS word
+    see /randomdox_generatot/README.md if needed to generate test .docx files
+
 ## recommanded to run test in python virtual environment, preventing python module mess up
 
 ## setup python virtual environment
@@ -48,16 +45,6 @@ The way the program works is it open MS word, generate random content and save i
 It will then revise the .docx for random(1 -3) times to let MS word generate more rsid
 Can use 'e' option to generate a new .docx form original .docx to obervse the change of rsid after a few edits of a .docx
 
-usage: $ python gen.py <option>
-
-options:
-    g <number of .docx>                     -----   generate <number of .docx> new .docx file in /testdocs
-    e <input .docx> <output .docx>          -----   randomly edit <input .docx> and save the result to newly created <output .docx> in /testdocs
-    rsid <document.docx>                    -----   extract and print rsid of <document.docx> in /testdocs
-    rsid_all                                -----   extract and print rsid details of  .docx in /testdocs, only works went filename is document_<int>.docx
-    rsid_sim <1.docx> <2.docx>              -----   calculate similarities between <1.docx> and <2.docx>
-    clean                                   -----   delete all contents in /testdocs
-
 log.py create logger object
 
 code example :
@@ -78,4 +65,4 @@ code example :
     logger.clean(<logs_file>)
 
 # performmance test
-    locust -f performancetest_remote.py --host=http://elijahmullens.com
+    locust -f performancetest/performancetest_remote.py --host=http://127.0.0.1:5000
